requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic-ai>=0.2.18
google-generativeai>=0.8.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
python-telegram-bot>=22.1